Merge branch 'wed_170_release' into feat_tyler_mode

# Conflicts:
#	SceneryTools_xcode6.xcodeproj/project.pbxproj
#	src/Utils/GISUtils.cpp
#	src/WEDCore/README.WorldEditor
#	src/WEDCore/WED_Validate.cpp
#	src/WEDCore/WED_Validate.h
#	src/WEDCore/WED_Version.h
#	src/WEDImportExport/WED_SceneryPackExport.cpp
#	src/WEDWindows/WED_GroupCommands.cpp
#	src/WEDWindows/WED_GroupCommands.h

+++ src/OneOffs/shape2xon.cpp
@@ -69,7 +69,7 @@ struct pt_t {
 	}
 	
 	void print(void) const {
-		printf("(%lf,%lf [0x%016llx,0x%016llx] ",x,y,x,y);
+		printf("(%lf,%lf [0x%016llx,0x%016llx] ",x,y,(long long) x, (long long) y);
 	}
 };
